% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breastfeeding.R
\name{breastfeeding}
\alias{breastfeeding}
\title{"My Breast Friend": nursing and diaper monitoring app}
\usage{
breastfeeding()
}
\value{
Launches a \code{Shiny} app that allows you to log new breastfeeding times and
dirty diapers and review your childcare history. Impress your pediatrician with solid data on
your baby's in's and out's! The app will look for a \code{data.csv} file within your
working directory, to which it will save new events and from which it will compile
user-history visualizations. If this file is not found, it will be created.
}
\description{
"My Breast Friend": nursing and diaper monitoring app
}
